+++ apps/native-component-list/src/navigation/ExpoApisStackNavigator.tsx
@@ -13,6 +13,7 @@ import { CalendarsNextScreens } from '../screens/CalendarsNextScreen';
 import { CalendarsScreens } from '../screens/CalendarsScreen';
 import { apiScreensToListElements } from '../screens/ComponentListScreen';
 import { ContactsScreens } from '../screens/Contacts/ContactsScreen';
+import { CryptoScreens } from '../screens/Crypto/CryptoScreen';
 import ExpoApis from '../screens/ExpoApisScreen';
 import { MediaLibraryScreens } from '../screens/MediaLibrary@Next/MediaLibraryScreens';
 import { ModulesCoreScreens } from '../screens/ModulesCore/ModulesCoreScreen';
@@ -305,7 +306,7 @@ export const ScreensList: ScreenConfig[] = [
   },
   {
     getComponent() {
-      return optionalRequire(() => require('../screens/CryptoScreen'));
+      return optionalRequire(() => require('../screens/Crypto/CryptoScreen'));
     },
     name: 'Crypto',
   },
@@ -474,6 +475,7 @@ export const Screens: ScreenConfig[] = [
   ...ContactsScreens,
   ...CalendarsScreens,
   ...CalendarsNextScreens,
+  ...CryptoScreens,
 ];
 
 export const screenApiItems = apiScreensToListElements(ScreensList);