+++ packages/expo-modules-core/ios/Uuidv5/Uuidv5.swift
@@ -12,7 +12,7 @@ func uuidv5(name: String, namespace: UUID) -> UUID {
 
   // Compute SHA1 digest
   var digest = [UInt8](repeating: 0, count: Int(CC_SHA1_DIGEST_LENGTH))
-  data.withUnsafeBytes { (ptr: UnsafeRawBufferPointer) -> Void in CC_SHA1(ptr.baseAddress, CC_LONG(data.count), &digest) }
+  data.withUnsafeBytes { (ptr: UnsafeRawBufferPointer) in CC_SHA1(ptr.baseAddress, CC_LONG(data.count), &digest) }
 
   // Set version bits:
   digest[6] = digest[6] & 0x0F | UInt8(5) << 4