+++ packages/expo-notifications/src/TopicSubscriptionModule.android.ts
@@ -1,0 +1,5 @@
+import { requireNativeModule } from 'expo-modules-core';
+
+import type { TopicSubscriptionModule } from './TopicSubscriptionModule.types';
+
+export default requireNativeModule<TopicSubscriptionModule>('ExpoTopicSubscriptionModule');