+++ packages/expo-router/plugin/src/index.ts
@@ -23,16 +23,10 @@ const withExpoHeadIos: ConfigPlugin = (config) => {
   });
 };
 
-const withGammaScreens: ConfigPlugin<
-  {
-    /** Enable experimental data loader support. Requires `web.output: 'static'` to be set in app config. */
-    unstable_splitView?: boolean;
-  } | void
-> = (config, props) => {
-  const value = props?.unstable_splitView ?? false;
+const withGammaScreens: ConfigPlugin = (config) => {
   return withPodfile(config, (config) => {
     if (!config.modResults.contents.includes('RNS_GAMMA_ENABLED')) {
-      config.modResults.contents = `ENV['RNS_GAMMA_ENABLED']='${value ? 1 : 0}'\n${config.modResults.contents}`;
+      config.modResults.contents = `# Set by expo-router. This enables Fabric-only features from react-native-screens\nENV['RNS_GAMMA_ENABLED'] ||= '1'\n${config.modResults.contents}`;
     }
     return config;
   });
@@ -58,13 +52,15 @@ const withRouter: ConfigPlugin<
     unstable_useServerMiddleware?: boolean;
     /** Enable experimental data loader support. Requires `web.output: 'static'` to be set in app config. */
     unstable_useServerDataLoaders?: boolean;
+    /** Enable experimental server-side rendering. When enabled with `web.output: 'server'`, HTML is rendered at request time instead of being pre-rendered at build time. */
+    unstable_useServerRendering?: boolean;
   } | void
 > = (config, _props) => {
   const props = _props || {};
   validate(schema, props);
 
   withExpoHeadIos(config);
-  withGammaScreens(config, props);
+  withGammaScreens(config);
 
   return {
     ...config,