+++ packages/expo-router/src/toolbar/index.ts
@@ -4,6 +4,7 @@ import {
   ToolbarMenuAction,
   ToolbarButton,
   ToolbarSpacer,
+  ToolbarSearchBarPreferredSlot,
   ToolbarView,
 } from './elements';
 
@@ -21,8 +22,8 @@ import {
  *       <Toolbar>
  *         <Toolbar.Spacer />
  *         <Toolbar.Button icon="magnifyingglass" tintColor={Color.ios.placeholderText} />
- *         <Toolbar.View style={{ width: 200 }}>
- *           <TextInput placeholder="Search" />
+ *         <Toolbar.View>
+ *           <TextInput style={{ width: 200 }} placeholder="Search" />
  *         </Toolbar.View>
  *         <Toolbar.Menu icon="ellipsis">
  *           <Toolbar.MenuAction icon="mail" title="Send email" onPress={() => {}} />
@@ -42,10 +43,18 @@ export const Toolbar = Object.assign(ToolbarHost, {
   MenuAction: ToolbarMenuAction,
   Button: ToolbarButton,
   Spacer: ToolbarSpacer,
+  SearchBarPreferredSlot: ToolbarSearchBarPreferredSlot,
   View: ToolbarView,
 });
 
-export { ToolbarMenu, ToolbarMenuAction, ToolbarButton, ToolbarSpacer, ToolbarView };
+export {
+  ToolbarMenu,
+  ToolbarMenuAction,
+  ToolbarButton,
+  ToolbarSpacer,
+  ToolbarSearchBarPreferredSlot,
+  ToolbarView,
+};
 
 export type {
   ToolbarProps,
@@ -53,5 +62,6 @@ export type {
   ToolbarMenuActionProps,
   ToolbarButtonProps,
   ToolbarSpacerProps,
+  ToolbarSearchBarPreferredSlotProps,
   ToolbarViewProps as ToolbarCustomViewProps,
 } from './elements';