+++ packages/expo-ui/ios/TextView.swift
@@ -5,11 +5,6 @@ import ExpoModulesCore
 
 public final class TextViewProps: UIBaseViewProps {
   @Field var text: String = ""
-  @Field var weight: FontWeight?
-  @Field var design: FontDesign?
-  @Field var size: Double?
-  @Field var lineLimit: Int?
-  @Field var color: Color?
 
   // Override default frame alignment for text views
   override var defaultFrameAlignment: Alignment { .leading }
@@ -23,18 +18,20 @@ public struct TextView: ExpoSwiftUI.View {
   }
 
   public var body: some View {
-    let hasDeprecatedFontProps = props.weight != nil || props.design != nil || props.size != nil
+    buildText(applyModifiers: false)
+      .applyModifiers(props.modifiers, appContext: props.appContext, globalEventDispatcher: props.globalEventDispatcher)
+  }
 
-    Text(props.text)
-      .if(hasDeprecatedFontProps) { text in
-        // TODO: remove this block of code once we remove the deprecated font props
-        text.font(.system(
-          size: CGFloat(props.size ?? 17),
-          weight: props.weight?.toSwiftUI() ?? .regular,
-          design: props.design?.toSwiftUI() ?? .default
-        ))
-      }
-      .lineLimit(props.lineLimit)
-      .foregroundColor(props.color)
+  internal func buildText(applyModifiers: Bool = true) -> Text {
+    var result = applyModifiers
+    ? Text(props.text).applyTextModifiers(props.modifiers, appContext: props.appContext)
+    : Text(props.text)
+
+    if let children = props.children {
+      result = children
+        .compactMap { ($0.childView as? TextView)?.buildText(applyModifiers: true) }
+        .reduce(result, +)
+    }
+    return result
   }
 }