+++ packages/expo-crypto/CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ### 🎉 New features
 
+- Added support for AES-GCM encryption. ([#41249](https://github.com/expo/expo/pull/41249) by [@barthap](https://github.com/barthap))
+
 ### 🐛 Bug fixes
+
+- [Android] Fix memory corruption when offset typed array was provided to `digest()` or `getRandomValues()`. ([#42186](https://github.com/expo/expo/pull/42186) by [@barthap](https://github.com/barthap))
 
 ### 💡 Others